+++ src/main/java/com/revature/ePort/auctionshowing/dtos/responses/UserAuctions.java
@@ -1,19 +1,20 @@
 package com.revature.ePort.auctionshowing.dtos.responses;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public class UserAuctions {
     private String title;
-    private int buyout_bid;
-    private int starting_bid;
+    private BigDecimal buyout_bid;
+    private BigDecimal starting_bid;
     private Timestamp startDate;
     private Timestamp expirationDate;
-    private int numberOfBids;
+    private BigDecimal numberOfBids;
 
     public UserAuctions() {
     }
 
-    public UserAuctions(int buyout_bid, int startingBid, Timestamp startDate, Timestamp expirationDate, String title, int numberOfBids) {
+    public UserAuctions(BigDecimal buyout_bid, BigDecimal startingBid, Timestamp startDate, Timestamp expirationDate, String title, BigDecimal numberOfBids) {
         this.buyout_bid = buyout_bid;
         this.starting_bid = startingBid;
         this.startDate = startDate;
@@ -22,19 +23,19 @@ public class UserAuctions {
         this.numberOfBids = numberOfBids;
     }
 
-    public double getBuyout_bid() {
+    public BigDecimal getBuyout_bid() {
         return buyout_bid;
     }
 
-    public void setBuyout_bid(int buyout_bid) {
+    public void setBuyout_bid(BigDecimal buyout_bid) {
         this.buyout_bid = buyout_bid;
     }
 
-    public double getStartingBid() {
+    public BigDecimal getStartingBid() {
         return starting_bid;
     }
 
-    public void setStartingBid(int startingBid) {
+    public void setStartingBid(BigDecimal startingBid) {
         this.starting_bid = startingBid;
     }
 
@@ -62,11 +63,11 @@ public class UserAuctions {
         this.title = title;
     }
 
-    public int getNumberOfBids() {
+    public BigDecimal getNumberOfBids() {
         return numberOfBids;
     }
 
-    public void setNumberOfBids(int numberOfBids) {
+    public void setNumberOfBids(BigDecimal numberOfBids) {
         this.numberOfBids = numberOfBids;
     }
