+++ src/main/java/com/revature/ePort/bid/BidRepository.java
@@ -9,17 +9,18 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.event.TransactionalEventListener;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Transactional
 public interface BidRepository extends CrudRepository<Bid, String> {
 
     @Query(value = "insert into bid values(?1,?2,?3,true)", nativeQuery = true)
-    void addNewBid(String auctionID, String userID, int amount);
+    void addNewBid(String auctionID, String userID, BigDecimal amount);
 
     @Modifying
     @Query(value = "update bid set amount = ?1 where user_id = ?2 and auction_showing_id = ?3",nativeQuery = true)
-    void updateBid(int amount, String userId, String auctionId);
+    void updateBid(BigDecimal amount, String userId, String auctionId);
 
     @Query(value = "select * from bid where user_id =?1",nativeQuery = true)
     List<Bid> bidHistory(String userID);
@@ -28,7 +29,7 @@ public interface BidRepository extends CrudRepository<Bid, String> {
     String auctionExists(String auctionShowingID, String userID);
 
     @Query(value = "select max(amount) from bid where auction_showing_id=?1",nativeQuery = true)
-    int maxAmount(String auctionID);
+    BigDecimal maxAmount(String auctionID);
 
 
     /*List<Bid> findByIdAuctionShowingID(String auctionShowingID);
